Clean up of the two error pages and config file.
The error pages now use the full-page layout and the standard page-view.
(If the standard page view does not go, then we are in serious trouble.) I
have also cut down "not_found.pt" so it has almost no external
dependencies, such as the menus and footer, as one of these may be the
objects that is not found. The only dependency that is left is on the
utility links, as there is no slot to over-ride these.

In addition I have removed more dojo stuff from the config.

--HG--
branch : GSContent
extra : convert_revision : svn%3A01e103ed-f575-20b1-d296-6440ccab0075/GSContent%401530